# directmeds/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Direct Meds Pharmacy API",
        "description": "HIPAA-compliant pharmacy management system API for Direct Meds. This API provides secure access to prescription management, product catalog, payment processing, and compliance features.",
        "contact": {
            "name": "Direct Meds API Support",
            "email": "api-support@directmeds.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "Direct Meds API Server"
        }
    ],
    "paths": {
...
        "/auth/register": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Register new user",
                "description": "Create a new user account and receive an API token",
                "operationId": "registerUser",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password",
                                    "password_confirmation",
                                    "user_type",
                                    "terms_accepted",
                                    "privacy_accepted"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "maxLength": 255,
                                        "example": "John Doe"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "maxLength": 255,
                                        "example": "john@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "minLength": 8,
                                        "example": "SecurePassword123!"
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "SecurePassword123!"
                                    },
                                    "user_type": {
                                        "type": "string",
                                        "enum": [
                                            "patient",
                                            "pharmacist",
                                            "prescriber"
                                        ],
                                        "example": "patient"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "pattern": "^[\\+]?[1-9][\\d]{0,15}$",
                                        "example": "+1234567890"
                                    },
                                    "date_of_birth": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "1990-01-15"
                                    },
                                    "license_number": {
                                        "type": "string",
                                        "maxLength": 255,
                                        "example": "LIC123456"
                                    },
                                    "license_state": {
                                        "type": "string",
                                        "maxLength": 2,
                                        "example": "CA"
                                    },
                                    "license_expiry": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-12-31"
                                    },
                                    "dea_number": {
                                        "type": "string",
                                        "maxLength": 255,
                                        "example": "DEA123456"
                                    },
                                    "npi_number": {
                                        "type": "string",
                                        "maxLength": 255,
                                        "example": "1234567890"
                                    },
                                    "first_name": {
                                        "type": "string",
                                        "maxLength": 255,
                                        "example": "John"
                                    },
                                    "last_name": {
                                        "type": "string",
                                        "maxLength": 255,
                                        "example": "Doe"
                                    },
                                    "address_line_1": {
                                        "type": "string",
                                        "maxLength": 255,
                                        "example": "123 Main St"
                                    },
                                    "city": {
                                        "type": "string",
                                        "maxLength": 255,
                                        "example": "San Francisco"
                                    },
                                    "state": {
                                        "type": "string",
                                        "maxLength": 255,
                                        "example": "California"
                                    },
                                    "postal_code": {
                                        "type": "string",
                                        "maxLength": 10,
                                        "example": "94102"
                                    },
                                    "terms_accepted": {
                                        "type": "boolean",
                                        "example": true
                                    },
                                    "privacy_accepted": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "User registered successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "User registered successfully"
                                        },
                                        "user": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "John Doe"
                                                },
                                                "email": {
                                                    "type": "string",
                                                    "example": "john@example.com"
                                                },
                                                "user_type": {
                                                    "type": "string",
                                                    "example": "patient"
                                                },
                                                "is_active": {
                                                    "type": "boolean",
                                                    "example": true
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "format": "date-time"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "token": {
                                            "type": "string",
                                            "example": "1|abc123def456ghi789..."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        },
                                        "errors": {
                                            "properties": {
                                                "email": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "The email has already been taken."
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Use Sanctum bearer token for authentication. Include the token in the Authorization header as 'Bearer {token}'.",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}